% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyQualityWeights.R
\name{applyQualityWeights}
\alias{applyQualityWeights}
\title{Apply pre-specified sample weights}
\usage{
applyQualityWeights(vobj, weights)
}
\arguments{
\item{vobj}{\code{EList} from \code{voom} or \code{voomWithDreamWeights}.}

\item{weights}{sample level weights}
}
\description{
Apply pre-specified sample weights by scaling existing precision weights
}
\details{
Apply pre-specified sample-level weights to the existing precision weights estimated from the data.  While the \code{limma::voomWithQualityWeights} function of Lui et al. (2015) estimates the sample-level weights from \code{voom} fit, here the weights are fixed beforehand.
}
\references{
{
  \insertRef{liu2015weight}{variancePartition}
}
}
\seealso{
\code{limma::voomWithQualityWeights}
}
