% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{BIC.MArrayLM}
\alias{BIC.MArrayLM}
\alias{BIC.MArrayLM2}
\title{BIC from model fit}
\usage{
\method{BIC}{MArrayLM}(object, vobj, ...)

\method{BIC}{MArrayLM2}(object, vobj, ...)
}
\arguments{
\item{object}{result of \code{dream()}}

\item{vobj}{\code{EList} used to fit model}

\item{...}{See \code{?stats::BIC}}
}
\description{
BIC from model fit

BIC from model fit using edf
}
