% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVarPart2.R
\name{calcVarPart}
\alias{calcVarPart}
\title{Compute variance statistics}
\usage{
calcVarPart(fit, ...)
}
\arguments{
\item{fit}{regression model}

\item{...}{other arguments passed to \code{fastglmm::varpart()}}
}
\description{
Compute fraction of variation attributable to each variable in regression model.  Also interpretable as the intra-class correlation after correcting for all other variables in the model.
}
\details{
this is now an iterface to \code{fastglmm::varpart()}
}
\examples{
library(lme4)
data(varPartData)

# Linear mixed model
fit <- lmer(geneExpr[1, ] ~ (1 | Tissue) + Age, info)
calcVarPart(fit)

# Linear model
# Note that the two models produce slightly different results
# This is expected: they are different statistical estimates
# of the same underlying value
fit <- lm(geneExpr[1, ] ~ Tissue + Age, info)
calcVarPart(fit)
}
\seealso{
\code{fastglmm::varpart()}
}
