% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{change_t_df}
\alias{change_t_df}
\title{change_t_df}
\usage{
change_t_df(t, df, d_target)
}
\arguments{
\item{t}{t-statistic}

\item{df}{current degrees of freedom}

\item{d_target}{target degrees of freedom}
}
\value{
Given a t-statistic with degrees of freedom df, return a new t-statistic corresponding to a target degrees of freedom.  Both t-statistics give the same p-value when compared to their respective degrees of freedom
}
\description{
change_t_df
}

