% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{classifyTestsF}
\alias{classifyTestsF}
\title{Multiple Testing Genewise Across Contrasts}
\usage{
classifyTestsF(object, ...)
}
\arguments{
\item{object}{numeric matrix of t-statistics or an 'MArrayLM2' object from which the t-statistics may be extracted.}

\item{...}{additional arguments}
}
\description{
For each gene, classify a series of related t-statistics as up, down or not significant.
}
\details{
Works like limma::classifyTestsF, except object can have a list of covariance matrices object$cov.coefficients.list, instead of just one in object$cov.coefficients
}
\seealso{
\code{limma::classifyTestsF}
}
