% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variancePartition.R
\name{colinearityScore}
\alias{colinearityScore}
\title{Colinearity score}
\usage{
colinearityScore(fit)
}
\arguments{
\item{fit}{regression model fit from lm() or lmer()}
}
\value{
returns the colinearity score between 0 and 1, where a score > 0.99 means the degree of colinearity is too high
}
\description{
Colinearity score for a regression model indicating if variables are too highly correlated to give meaningful results
}
\examples{
# load library
# library(variancePartition)

# load simulated data:
data(varPartData)
form = ~ Age + (1|Individual) + (1|Tissue)

res = fitVarPartModel( geneExpr[1:10,], form, info )

# evaluate the colinearity score on the first model fit
# this reports the correlation matrix between coefficients estimates
# for fixed effects
# the colinearity score is the maximum absolute correlation value
# If the colinearity score > .99 then the variance parition
# estimates may be problematic
# In that case, a least one variable should be omitted
colinearityScore(res[[1]])
}

