% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{.getAllUniContrasts}
\alias{.getAllUniContrasts}
\title{Get all univariate contrasts}
\usage{
.getAllUniContrasts(exprObj, formula, data)
}
\arguments{
\item{exprObj}{matrix of expression data (g genes x n samples), or ExpressionSet, or EList returned by voom() from the limma package}

\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used a a response. e.g.: ~ a + b + (1|c)  Formulas with only fixed effects also work}

\item{data}{data.frame with columns corresponding to formula}
}
\value{
Matrix testing each variable one at a time.  Contrasts are on rows
}
\description{
Get all univariate contrasts
}
