% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.isMixedModelFormula}
\alias{.isMixedModelFormula}
\title{Check if model contains a random effect}
\usage{
.isMixedModelFormula(formula, data)
}
\arguments{
\item{formula}{model formula}

\item{data}{data.frame}
}
\description{
Check if model contains a random effect
}
