% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_functions.R
\name{dscchisq}
\alias{dscchisq}
\title{Scaled chi-square}
\usage{
dscchisq(x, a, b)
}
\arguments{
\item{x}{vector of quantiles.}

\item{a}{scale}

\item{b}{degrees of freedom}
}
\description{
Scaled chi-square density using a gamma distribution
}
