% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{eBayes,MArrayLM-method}
\alias{eBayes,MArrayLM-method}
\title{eBayes for MArrayLM}
\usage{
\S4method{eBayes}{MArrayLM}(fit, ...)
}
\arguments{
\item{fit}{fit}

\item{...}{all other args}
}
\description{
eBayes for result of linear  model for with uses \code{limma::eBayes()}
}
