% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{eBayes}
\alias{eBayes}
\alias{eBayes,MArrayLM-method}
\alias{eBayes,MArrayLM2-method}
\title{eBayes generic for for MArrayLM and MArrayLM2}
\usage{
eBayes(fit, ...)

\S4method{eBayes}{MArrayLM}(fit, ...)

\S4method{eBayes}{MArrayLM2}(
  fit,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  span = NULL,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1),
  legacy = NULL
)
}
\arguments{
\item{fit}{fit}

\item{...}{all other args}

\item{proportion}{proportion}

\item{stdev.coef.lim}{stdev.coef.lim}

\item{trend}{trend}

\item{span}{span}

\item{robust}{robust}

\item{winsor.tail.p}{winsor.tail.p}

\item{legacy}{legacy}
}
\value{
results of eBayes using approximated residual degrees of freedom
}
\description{
eBayes for result of linear (mixed) model

eBayes for result of linear  model for with uses \code{limma::eBayes()}

eBayes for result of linear mixed model for with \code{dream()} using residual degrees of freedom approximated with \code{rdf.merMod()}
}
\seealso{
\code{dream()}, \code{rdf.merMod()}, \code{limma::eBayes()}
}
