% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{eBayes}
\alias{eBayes}
\title{eBayes generic for for MArrayLM and MArrayLM2}
\usage{
eBayes(fit, ...)
}
\arguments{
\item{fit}{fit}

\item{...}{all other args}
}
\description{
eBayes for result of linear (mixed) model
}
