% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\docType{methods}
\name{eval_lmm}
\alias{eval_lmm}
\title{Evaluate contrasts for linear mixed model}
\usage{
eval_lmm(fit, L, ddf)
}
\arguments{
\item{fit}{model fit}

\item{L}{contrast matrix}

\item{ddf}{Specifiy "Satterthwaite" or "Kenward-Roger" method to estimate effective degress of freedom for hypothesis testing in the linear mixed model.  Note that Kenward-Roger is more accurate, but is *much* slower.  Satterthwaite is a good enough exproximation for most datasets.}
}
\value{
df, sigma, beta, SE of model
}
\description{
Evaluate contrasts for linear mixed model
}

