% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMT.R
\name{FMT}
\alias{FMT}
\title{Fully moderated t-test}
\usage{
FMT(
  Amean,
  sigmasq,
  df,
  span1 = 0.5,
  span2 = 0.95,
  iter1 = 4,
  iter2 = 4,
  b = 20
)
}
\arguments{
\item{Amean}{average log intensity levels of all genes}

\item{sigmasq}{residual variances of all genes}

\item{df}{degrees of freedom for sigmasq}

\item{span1}{span parameter in LOESS smoothing function}

\item{span2}{span parameter in LOESS smoothing function}

\item{iter1}{iteration number in LOESS smoothing function}

\item{iter2}{iteration number in LOESS smoothing function}

\item{b}{number of genes on either side of moving average window when calculating variance of log residual variances}
}
\description{
Fully moderated t-test
}
\keyword{internal}
