% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMT.R
\name{FMT.ZI}
\alias{FMT.ZI}
\title{Fully moderated t-test for zero-inflated variances}
\usage{
FMT.ZI(
  Amean,
  sigmasq,
  df,
  span1 = 0.5,
  span2 = 0.95,
  iter1 = 4,
  iter2 = 4,
  b = 20,
  sigma.thres = 1e-07
)
}
\arguments{
\item{Amean}{average log intensity levels of all genes}

\item{sigmasq}{residual variances of all genes}

\item{df}{degrees of freedom for variance component}

\item{span1}{span parameter in LOESS smoothing function}

\item{span2}{span parameter in LOESS smoothing function}

\item{iter1}{iteration number in LOESS smoothing function}

\item{iter2}{iteration number in LOESS smoothing function}

\item{b}{number of genes on either side of moving average window when calculating variance of log residual variances}

\item{sigma.thres}{threshold for variance being zero}
}
\description{
Fully moderated t-test for zero-inflated variances
}
\keyword{internal}
