% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prediction.R
\name{get_prediction,lm-method}
\alias{get_prediction,lm-method}
\title{Compute predicted value of formula for linear model}
\usage{
\S4method{get_prediction}{lm}(fit, formula)
}
\arguments{
\item{fit}{model fit with \code{lm}}

\item{formula}{formula ofeffects to predict}
}
\value{
Predicted values from formula using parameter estimates from fit linear model
}
\description{
Compute predicted value of formula for linear model for with \code{lm}
}
\examples{

library(lme4)

# fit model
fit <- lm(Reaction ~ Days, sleepstudy)

# prediction of intercept
get_prediction( fit, ~ 1)

# prediction of Days without intercept
get_prediction( fit, ~ 0 + Days)

}
