% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{get_standardized_t}
\alias{get_standardized_t}
\title{Convert t-statistics to standard degrees of freedom}
\usage{
get_standardized_t(fit, coef = NULL, number = 10)
}
\arguments{
\item{fit}{fit from dream}

\item{coef}{coef passed to topTable}

\item{number}{number passed to TopTable}
}
\description{
Convert t-statistics with differing degrees of freedom (df) to standard degrees of freedom.
}
\examples{
# load library
# library(variancePartition)
library(BiocParallel)

# load simulated data:
# geneExpr: matrix of gene expression values
# info: information/metadata about each sample
data(varPartData)

form <- ~ Batch + (1|Individual) + (1|Tissue) 

# Fit linear mixed model for each gene
# run on just 10 genes for time
fit = dream( geneExpr[1:10,], form, info)

get_standardized_t(fit, "Batch2")

}
