% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{ggColorHue}
\alias{ggColorHue}
\title{Default colors for ggplot}
\usage{
ggColorHue(n)
}
\arguments{
\item{n}{number of colors}
}
\value{
array of colors of length n
}
\description{
Return an array of n colors the same as the default used by ggplot2
}
\examples{
ggColorHue(4)
}

