% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatvalues.R
\name{hatvalues,MArrayLM-method}
\alias{hatvalues,MArrayLM-method}
\alias{hatvalues,MArrayLM2-method}
\title{Compute hatvalues}
\usage{
\S4method{hatvalues}{MArrayLM}(model, vobj, ...)

\S4method{hatvalues}{MArrayLM2}(model, ...)
}
\arguments{
\item{model}{model fit from \code{dream()}}

\item{vobj}{\code{EList} returned by \code{voom()} or \code{voomWithDreamWeights()}.}

\item{...}{other arguments, currently ignored}
}
\description{
Compute hatvalues from dream fit
}
