% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isRunableFormula.R
\name{isRunableFormula}
\alias{isRunableFormula}
\title{Test if formula is full rank on this dataset}
\usage{
isRunableFormula(exprObj, formula, data)
}
\arguments{
\item{exprObj}{expression object}

\item{formula}{formula}

\item{data}{data}
}
\description{
Test if formula is full rank on this dataset
}
