% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.R
\name{logLik.MArrayLM}
\alias{logLik.MArrayLM}
\alias{logLik.MArrayLM2}
\title{Log-likelihood from model fit}
\usage{
\method{logLik}{MArrayLM}(object, vobj, ...)

\method{logLik}{MArrayLM2}(object, ...)
}
\arguments{
\item{object}{result of \code{lmFit()} or \code{dream()}}

\item{vobj}{\code{EList} used to fit model}

\item{...}{See \code{?stats::logLik}}
}
\description{
Log-likelihood from model fit

Log-likelihood from model fit
}
