% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPercentBars}
\alias{plotPercentBars}
\title{Bar plot of variance fractions}
\usage{
plotPercentBars(varPart, col = c(ggColorHue(ncol(varPart) - 1), "#bebebe99"))
}
\arguments{
\item{varPart}{object returned by extractVarPart() or fitExtractVarPartModel()}

\item{col}{color of bars for each variable}
}
\value{
Returns ggplot2 barplot
}
\description{
Bar plot of variance fractions for a subset of genes
}
\examples{
# library(variancePartition)

# optional step to run analysis in parallel on multicore machines
# Here, we used 4 threads
library(doParallel)
cl <- makeCluster(4)
registerDoParallel(cl)
# or by using the doSNOW package

# load simulated data:
# geneExpr: matrix of gene expression values
# info: information/metadata about each sample
data(varPartData)

# Specify variables to consider
form <- ~ Age + (1|Individual) + (1|Tissue)

# Fit model
varPart <- fitExtractVarPartModel( geneExpr, form, info )

# Bar plot for a subset of genes showing variance fractions
plotPercentBars( varPart[1:5,] )

# Move the legend to the top
plotPercentBars( varPart[1:5,] ) + theme(legend.position="top") 

# stop cluster
stopCluster(cl)
}

