% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_functions.R
\name{plotVarianceEstimates}
\alias{plotVarianceEstimates}
\title{Plot Variance Estimates}
\usage{
plotVarianceEstimates(
  fit,
  fitEB,
  var_true = NULL,
  xmax = quantile(fit$sigma^2, 0.999)
)
}
\arguments{
\item{fit}{model fit from \code{dream()}}

\item{fitEB}{model fit from \code{eBayes()}}

\item{var_true}{array of true variance values from simulation (optional)}

\item{xmax}{maximum value on the x-axis}
}
\description{
Plot Variance Estimates
}
