% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_functions.R
\name{rdf}
\alias{rdf}
\title{Residual degrees of freedom}
\usage{
rdf(fit)
}
\arguments{
\item{fit}{model fit from \code{lm()}, \code{glm()}, \code{lmer()}}
}
\description{
Residual degrees of freedom
}
\examples{
library(lme4)

fit <- lm(Reaction ~ Days, sleepstudy)
rdf(fit)

}
\seealso{
\code{rdf.merMod}
}
