% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf_functions.R
\name{rdf_from_matrices}
\alias{rdf_from_matrices}
\title{Fast approximate residual degrees of freedom}
\usage{
rdf_from_matrices(A, B)
}
\arguments{
\item{A}{a \code{matrix} or \code{sparseMatrix}}

\item{B}{a \code{matrix} or \code{sparseMatrix}}
}
\description{
Defining \eqn{H = A^TA + B^TB} where \eqn{A} and \eqn{B} are low rank, compute
\eqn{n - 2tr(H) + tr(HH)} in \eqn{O(np^2)} instead of \eqn{O(n^2p^2)}.
}
\seealso{
rdf.merMod
}
