% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{reOnly}
\alias{reOnly}
\title{Adapted from lme4:::reOnly}
\usage{
reOnly(f, response = FALSE)
}
\arguments{
\item{f}{formula}

\item{response}{(FALSE) is there a response in the formula}
}
\description{
Adapted from lme4:::reOnly
}
