% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{residuals,MArrayLM2-method}
\alias{residuals,MArrayLM2-method}
\title{residuals for MArrayLM2}
\usage{
\S4method{residuals}{MArrayLM2}(object, y, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{MArrayLM2 object from dream}

\item{y}{\code{EList} object used in \code{dream()}}

\item{type}{compute either response or pearson residuals}

\item{...}{other arguments, currently ignored}
}
\value{
results of residuals
}
\description{
residuals for MArrayLM2
}
