% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{residuals.MArrayLM2}
\alias{residuals.MArrayLM2}
\title{Residuals for result of dream}
\usage{
residuals.MArrayLM2(object, ...)
}
\arguments{
\item{object}{See \code{?stats::residuals}}

\item{...}{See \code{?stats::residuals}}
}
\description{
Residuals for result of dream
}
