% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{residuals.MArrayLM2}
\alias{residuals.MArrayLM2}
\alias{residuals,MArrayLM-method}
\alias{residuals,MArrayLM2-method}
\title{Residuals for result of dream}
\usage{
residuals.MArrayLM2(object, y, ..., type = c("response", "pearson"))

\S4method{residuals}{MArrayLM}(object, y, ..., type = c("response", "pearson"))

\S4method{residuals}{MArrayLM2}(object, y, type = c("response", "pearson"), ...)
}
\arguments{
\item{object}{MArrayLM2 object from dream}

\item{y}{\code{EList} object used in \code{dream()}}

\item{...}{other arguments, currently ignored}

\item{type}{compute either response or pearson residuals}
}
\value{
results of residuals

results of residuals
}
\description{
Residuals for result of dream

residuals for MArrayLM

residuals for MArrayLM2
}
