% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{[.MArrayLM2}
\alias{[.MArrayLM2}
\alias{subset.MArrayLM2,MArrayLM2-method}
\title{Subseting for MArrayLM2}
\arguments{
\item{object}{MArrayLM2}

\item{i}{row}

\item{j}{col}
}
\value{
subset
}
\description{
Enable subsetting on MArrayLM2 object. Same as for MArrayLM, but apply column subsetting to df.residual and cov.coefficients.list
}
\keyword{internal}
