% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotation.R
\docType{data}
\name{varPartData}
\alias{varPartData}
\alias{geneCounts}
\alias{info}
\alias{geneExpr}
\title{Simulation dataset for examples}
\format{
A dataset of 100 samples and 200 genes

A dataset of 100 samples and 200 genes

A dataset of 100 samples and 200 genes

A dataset of 100 samples and 200 genes
}
\usage{
data(varPartData)

data(varPartData)

data(varPartData)

data(varPartData)
}
\description{
A simulated dataset of gene expression and metadata

A simulated dataset of gene counts

A simulated dataset of gene counts

A simulated dataset of gene counts
}
\details{
\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}

\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}

\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}

\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}
}
\keyword{datasets}
