% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotation.R
\docType{data}
\name{varPartDEdata}
\alias{varPartDEdata}
\alias{countMatrix}
\alias{metadata}
\title{Simulation dataset for dream example}
\format{
A dataset of 24 samples and 19,364 genes

A dataset of 24 samples and 19,364 genes
}
\usage{
data(varPartDEdata)

data(varPartDEdata)
}
\description{
Gene counts from RNA-seq

metadata matrix of sample information
}
\details{
\itemize{    
\item countMatrix gene expression in the form of RNA-seq counts
 \item metadata metadata about the study design
}

\itemize{    
\item countMatrix gene expression in the form of RNA-seq counts
 \item metadata metadata about the study design
}
}
\keyword{datasets}
