% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov,MArrayLM2-method}
\alias{vcov,MArrayLM2-method}
\title{Co-variance matrix for \code{dream()} fit}
\usage{
\S4method{vcov}{MArrayLM2}(object, vobj, coef)
}
\arguments{
\item{object}{\code{MArrayLM} object return by \code{lmFit()} or \code{dream()}}

\item{vobj}{\code{EList} object returned by \code{voom()}}

\item{coef}{name of coefficient to be extracted}
}
\value{
variance-covariance matrix
}
\description{
Define generic \code{vcov()} for result of \code{lmFit()} and \code{dream()}
}
