% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/variancePartition.R
\docType{methods}
\name{calcVarPart}
\alias{calcVarPart}
\alias{calcVarPart,lm-method}
\alias{calcVarPart,lmerMod-method}
\title{Compute variance statistics}
\usage{
calcVarPart(fit, ...)

\S4method{calcVarPart}{lm}(fit, ...)

\S4method{calcVarPart}{lmerMod}(fit, ...)
}
\arguments{
\item{fit}{model fit from lm() or lmer()}

\item{...}{additional arguments (not currently used)}
}
\value{
fraction of variance explained / ICC for each variable in the model
}
\description{
Compute fraction of variation attributable to each variable in regression model.  Also interpretable as the intra-class correlation after correcting for all other variables in the model
}
\examples{
data(varPartData)

# Linear mixed model
fit = lmer( geneExpr[1,] ~ (1|Tissue) + Age, info)
calcVarPart( fit )

# Linear model
# Note that the two models produce slightly different results
# This is expected: they are different statistical estimates
# of the same underlying value
fit = lm( geneExpr[1,] ~ Tissue + Age, info)
calcVarPart( fit )
}

