% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{change_t_df}
\alias{change_t_df}
\title{Example dataset for dream}
\usage{
change_t_df(t, df, d_target)
}
\arguments{
\item{t}{t-statistic}

\item{df}{current degrees of freedom}

\item{d_target}{target degrees of freedom}
}
\value{
Given a t-statistic with degrees of freedom df, return a new t-statistic corresponding to a target degrees of freedom.  Both t-statistics give the same p-value when compared to their respective degrees of freedom
}
\description{
Simulated RNA-seq counts in 'countMatrix' and information in 'metadata'
}
\details{
change_t_df

change_t_df
}

