% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrasts.R
\name{.getAllUniContrasts}
\alias{.getAllUniContrasts}
\title{Get all univariate contrasts}
\usage{
.getAllUniContrasts(formula, data)
}
\arguments{
\item{formula}{specifies variables for the linear (mixed) model.  Must only specify covariates, since the rows of exprObj are automatically used as a response. e.g.: \code{~ a + b + (1|c)}  Formulas with only fixed effects also work}

\item{data}{data.frame with columns corresponding to formula}
}
\value{
Matrix testing each variable one at a time.  Contrasts are on rows
}
\description{
Get all univariate contrasts
}
\keyword{internal}
