% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{.isMixedModelFormula}
\alias{.isMixedModelFormula}
\title{Check if model contains a random effect}
\usage{
.isMixedModelFormula(formula)
}
\arguments{
\item{formula}{model formula}
}
\description{
Check if model contains a random effect
}
\keyword{internal}
