% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\docType{methods}
\name{eBayes,MArrayLMM_lmer-method}
\alias{eBayes,MArrayLMM_lmer-method}
\title{eBayes for MArrayLMM_lmer}
\usage{
\S4method{eBayes}{MArrayLMM_lmer}(fit, proportion = 0.01,
  stdev.coef.lim = c(0.1, 4), trend = FALSE, robust = FALSE,
  winsor.tail.p = c(0.05, 0.1))
}
\arguments{
\item{fit}{fit}

\item{proportion}{proportion}

\item{stdev.coef.lim}{stdev.coef.lim}

\item{trend}{trend}

\item{robust}{robust}

\item{winsor.tail.p}{winsor.tail.p}
}
\description{
eBayes for MArrayLMM_lmer
}

