% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\name{eBayes,MArrayLM2-method}
\alias{eBayes,MArrayLM2-method}
\title{eBayes for MArrayLM2}
\usage{
\S4method{eBayes}{MArrayLM2}(
  fit,
  proportion = 0.01,
  stdev.coef.lim = c(0.1, 4),
  trend = FALSE,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1)
)
}
\arguments{
\item{fit}{fit}

\item{proportion}{proportion}

\item{stdev.coef.lim}{stdev.coef.lim}

\item{trend}{trend}

\item{robust}{robust}

\item{winsor.tail.p}{winsor.tail.p}
}
\value{
results of eBayes using approximated residual degrees of freedom
}
\description{
eBayes for result of linear mixed model for with \code{dream()} using residual degrees of freedom approximated with \code{rdf.merMod()}
}
\seealso{
dream rdf.merMod
}
