% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.R
\name{plotTrendVP}
\alias{plotTrendVP}
\alias{plotTrendVP,DESeqDataSet,data.frame-method}
\alias{plotTrendVP,DGELRT,data.frame-method}
\title{Plot Trend of Variance Fractions Vs Count Magnitude}
\usage{
plotTrendVP(x, vp, component, ...)

\S4method{plotTrendVP}{DESeqDataSet,data.frame}(x, vp, component, ...)

\S4method{plotTrendVP}{DGELRT,data.frame}(x, vp, component, ...)
}
\arguments{
\item{x}{regression fits}

\item{vp}{\code{data.frame} from \code{fitVarPart()}}

\item{component}{variance component to extract from \code{vp}}

\item{...}{additional arguments}
}
\value{
Plot of variance fraction vs count magnitude
}
\description{
Plot trend of variance fractions for a specified component versus count magnitude for each gene and cell cluster
}
\examples{
# Simulate counts
set.seed(1)
countMatrix <- matrix(rnbinom(n=100000, mu=20, size=3), ncol=10)
rownames(countMatrix) <- paste0("gene_", seq(nrow(countMatrix)))
colnames(countMatrix) <- paste0("sample_", seq(ncol(countMatrix)))

condition <- factor(rep(1:2, each=5))

# DESeq2 model #
library(DESeq2)
dds <- DESeqDataSetFromMatrix(countMatrix, 
  DataFrame(condition), 
  ~ condition)
dds <- DESeq(dds, quiet=TRUE)
res <- results(dds)

# Variance partition analysis
vp1 <- varpart(dds)

# Plot count noise vs expression magnitude
plotTrendVP( dds, vp1, "CountNoise" )


# edgeR model #
library(edgeR)
design <- model.matrix( ~ condition, data.frame(condition))
d <- DGEList(countMatrix)
d <- normLibSizes(d)
d <- estimateDisp(d, design)
fit <- glmQLFit(d, design)
fit <- glmQLFTest(fit)

vp2 <- varpart(fit, dispObj = d, formula = ~ cond)

# Plot count noise vs expression magnitude
plotTrendVP( fit, vp2, "CountNoise" )

}
