% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitMixedModelDE.R
\docType{methods}
\name{toptable,MArrayLMM_lmer-method}
\alias{toptable,MArrayLMM_lmer-method}
\title{toptable for MArrayLMM_lmer}
\usage{
\S4method{toptable}{MArrayLMM_lmer}(fit, coef = 1, number = 10,
  genelist = NULL, A = NULL, eb = NULL, adjust.method = "BH",
  sort.by = "B", resort.by = NULL, p.value = 1, lfc = 0,
  confint = FALSE, ...)
}
\arguments{
\item{fit}{fit}

\item{coef}{coef}

\item{number}{number}

\item{genelist}{genelist}

\item{A}{A}

\item{eb}{eb}

\item{adjust.method}{adjust.method}

\item{sort.by}{sort.by}

\item{resort.by}{resort.by}

\item{p.value}{p.value}

\item{lfc}{lfc}

\item{confint}{confint}

\item{...}{...}
}
\value{
resold of toptable
}
\description{
toptable for MArrayLMM_lmer
}

