% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toptable.R
\docType{methods}
\name{topTable,MArrayLM2-method}
\alias{topTable,MArrayLM2-method}
\alias{toptable,MArrayLM2-method}
\title{toptable for MArrayLMM_lmer}
\usage{
\S4method{topTable}{MArrayLM2}(fit, coef = NULL, number = 10,
  genelist = fit$genes, adjust.method = "BH", sort.by = "p",
  resort.by = NULL, p.value = 1, lfc = 0, confint = FALSE)
}
\arguments{
\item{fit}{fit}

\item{coef}{coef}

\item{number}{number}

\item{genelist}{genelist}

\item{adjust.method}{adjust.method}

\item{sort.by}{sort.by}

\item{resort.by}{resort.by}

\item{p.value}{p.value}

\item{lfc}{lfc}

\item{confint}{confint}
}
\value{
results of toptable
}
\description{
toptable for MArrayLMM_lmer
}
