% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBayesFMT.R, R/toptable.R
\name{topTable,FMT-method}
\alias{topTable,FMT-method}
\alias{toptable,FMT-method}
\alias{topTable,MArrayLM2-method}
\alias{toptable,MArrayLM2-method}
\title{toptable for FMT}
\usage{
\S4method{topTable}{FMT}(
  fit,
  coef = NULL,
  number = 10,
  genelist = fit$genes,
  adjust.method = "BH",
  sort.by = "p",
  resort.by = NULL,
  p.value = 1,
  lfc = 0,
  confint = FALSE
)

\S4method{topTable}{MArrayLM2}(
  fit,
  coef = NULL,
  number = 10,
  genelist = fit$genes,
  adjust.method = "BH",
  sort.by = "p",
  resort.by = NULL,
  p.value = 1,
  lfc = 0,
  confint = FALSE
)
}
\arguments{
\item{fit}{fit}

\item{coef}{coef}

\item{number}{number}

\item{genelist}{genelist}

\item{adjust.method}{adjust.method}

\item{sort.by}{sort.by}

\item{resort.by}{resort.by}

\item{p.value}{p.value}

\item{lfc}{lfc}

\item{confint}{confint}
}
\value{
results of toptable

results of toptable
}
\description{
toptable for FMT

toptable for MArrayLM2
}
