% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varpart.R
\name{varpart,DESeqDataSet-method}
\alias{varpart,DESeqDataSet-method}
\alias{varpart,DGEGLM-method}
\alias{varpart,DGELRT-method}
\title{Variance Partitioning Analysis}
\usage{
\S4method{varpart}{DESeqDataSet}(
  fit,
  method = c("exact", "approximate"),
  pseudocount = 1,
  p.tail = 1e-04,
  nthreads = parallelly::availableCores(),
  ...
)

\S4method{varpart}{DGEGLM}(
  fit,
  method = c("exact", "approximate"),
  pseudocount = 1,
  p.tail = 1e-04,
  nthreads = parallelly::availableCores(),
  dispObj,
  formula,
  ...
)

\S4method{varpart}{DGELRT}(
  fit,
  method = c("exact", "approximate"),
  pseudocount = 1,
  p.tail = 1e-04,
  nthreads = parallelly::availableCores(),
  dispObj,
  formula,
  ...
)
}
\arguments{
\item{fit}{regression model fit}

\item{method}{select method for count models:  \code{"exact"} or \code{"approximate"} for faster approximation}

\item{pseudocount}{pseudocount used for \code{"exact"} and \code{"approximate"} methods for count models}

\item{p.tail}{probability threashold for evaluating expectations for \code{"exact"} methods for count models}

\item{nthreads}{number of threads used for count models}

\item{...}{other arguments}

\item{dispObj}{result of \code{estimateDisp()}}

\item{formula}{formula used for the design matrix}
}
\description{
Variance partitioning analysis on each gene from regression models fit with \code{edgeR} and \code{DESeq2}
}
\examples{
library(ggplot2)

# Simulate counts
set.seed(1)
eta <- rnorm(10, 3, 1)
mu <- exp(eta)

countMatrix <- matrix(rnbinom(n=100000, mu=mu, size=3), ncol=10)
rownames(countMatrix) <- paste0("gene_", seq(nrow(countMatrix)))
colnames(countMatrix) <- paste0("sample_", seq(ncol(countMatrix)))

condition <- factor(rep(1:2, each=5))

# DESeq2 model #
library(DESeq2)
dds <- DESeqDataSetFromMatrix(countMatrix, 
  DataFrame(condition), 
  ~ condition)
dds <- DESeq(dds, quiet=TRUE)
res <- results(dds)

# Variance partition analysis
vp1 <- varpart(dds)

# Plot contribution of each component
plotVarPart(vp1, main="DESeq2") + theme(aspect.ratio=1)

# Plot count noise vs expression magnitude
plotTrendVP( dds, vp1, "CountNoise" )

# edgeR model #
library(edgeR)
design <- model.matrix( ~ condition, data.frame(condition))
d <- DGEList(countMatrix)
d <- normLibSizes(d)
d <- estimateDisp(d, design)
fit <- glmQLFit(d, design)
fit <- glmQLFTest(fit)

vp2 <- varpart(fit, dispObj = d, formula = ~ cond)

# Plot contribution of each component
plotVarPart(vp2, main="edgeR") + theme(aspect.ratio=1)

# Plot count noise vs expression magnitude
plotTrendVP( fit, vp2, "CountNoise", dispObj = d )

}
\seealso{
\code{fastglmm::varpart()}
}
