% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variancePartition.R
\docType{data}
\name{varPartData}
\alias{varPartData}
\title{Simulation dataset for examples}
\format{A dataset of 100 samples and 200 genes}
\usage{
data(varPartData)
}
\description{
A simulated dataset of gene expression and metadata
}
\details{
\itemize{    
\item geneCounts: gene expression in the form of RNA-seq counts
\item geneExpr: gene expression on a continuous scale
 \item info: metadata about the study design
}
}
\keyword{datasets}

\alias{geneCounts}
\alias{geneExpr}
\alias{info}
\value{varPartData}
