% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotation.R
\docType{data}
\name{varPartDEdata}
\alias{varPartDEdata}
\alias{countMatrix}
\alias{metadata}
\title{A simulated dataset of gene counts}
\format{
A dataset of 24 samples and 19,364 genes

A dataset of 24 samples and 19,364 genes
}
\usage{
data(varPartData)

data(varPartData)
}
\description{
\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}

\itemize{
	\item geneCounts gene expression in the form of RNA-seq counts
	\item geneExpr gene expression on a continuous scale
 \item info metadata about the study design
}
}
\keyword{datasets}
